feat: decorator tests, integration in ui
