+++ src/mock_node/slider_values.txt
@@ -18,7 +18,7 @@
 19
 30
 30
-30
+11
 21
 30
 30
