+++ src/openadr_node/virtual_end_node.py
@@ -12,7 +12,6 @@ logger = logging.getLogger(__name__)
 
 class VirtualEndNode:
   def __init__(self, ven_name: str, vtn_url: str):
-    print("INIT VEN")
     self._ven_name = ven_name
     self._vtn_url = vtn_url
     self._open_adr_client = OpenADRClient(self._ven_name, self._vtn_url)
@@ -36,13 +35,13 @@ class VirtualEndNode:
   async def handle_event(event):
     # business logic
     # handle event
-    print("DISPATCHING ACTION GOTTEN")
+    print('DISPATCHING ACTION GOTTEN')
     logging.info('Processing openADR Event')
     _event_descriptor = event['event_descriptor']
     _active_period = event['active_period']
     _event_signals = event['event_signals']
     _targets = event['targets']
-    print("EVENT!!")
+    print('EVENT!!')
     return 'optIn'  # eventually pass devices status?
 
   def get_open_adr_server_run(self):
